getConnection() returns null if connection not found

+++ app/diograph-store.ts
@@ -94,11 +94,15 @@ export class DiographStore {
   static deleteConnection(fromDioryId, toDioryId): Promise<any> {
     if (fromDioryId === undefined || toDioryId === undefined) { throw "Required two ids not given to DiographStore.deleteConnection()" }
     return this.getConnection(fromDioryId, toDioryId).then(connection => {
-      return DiographApi.delete(connection.id, "connections").then(response => {
-        // let connectionDatastoreModel = this.datastore.find("connections", connection.id)
-        // this.datastore.destroy(connectionDatastoreModel)
+      if (connection === null) {
         return null
-      })
+      } else {
+        return DiographApi.delete(connection.id, "connections").then(response => {
+          // let connectionDatastoreModel = this.datastore.find("connections", connection.id)
+          // this.datastore.destroy(connectionDatastoreModel)
+          return null
+        })
+      }
     })
   }
